LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/microtasks/serviceworker.any.serviceworker-expected.txt


FAIL import() should not drain the microtask queue if it fails because it's used in a ServiceWorker assert_less_than: expected a number less than 1000 but got 1000
